fix: Clean up stories and tests
