+++ modules/preview-react/pill/stories/examples/WithAvatar.tsx
@@ -3,14 +3,20 @@ import React from 'react';
 import {Pill} from '@workday/canvas-kit-preview-react/pill';
 // @ts-ignore: Cannot find module error
 import testAvatar from './test-avatar.png';
-import {Flex, Box} from '@workday/canvas-kit-react/layout';
 import {BodyText} from '@workday/canvas-kit-react/text';
+import {createStyles} from '@workday/canvas-kit-styling';
+import {system} from '@workday/canvas-tokens-web';
+
+const flexStyles = createStyles({
+  display: 'flex',
+  gap: system.space.x2,
+});
 
 export const WithAvatar = () => {
   const [text, setText] = React.useState('');
   return (
-    <Box>
-      <Flex gap="xxs">
+    <div>
+      <div className={flexStyles}>
         <Pill onClick={() => setText('The first pill is clicked!')}>
           <Pill.Avatar url={testAvatar} />
           Regina Skeltor
@@ -19,8 +25,8 @@ export const WithAvatar = () => {
           <Pill.Avatar url={testAvatar} />
           Regina Skeltor
         </Pill>
-      </Flex>
+      </div>
       <BodyText size="medium">{text}</BodyText>
-    </Box>
+    </div>
   );
 };
