+++ modules/preview-react/pill/stories/examples/WithCount.tsx
@@ -1,13 +1,19 @@
 import React from 'react';
 import {Pill} from '@workday/canvas-kit-preview-react/pill';
-import {Flex, Box} from '@workday/canvas-kit-react/layout';
 import {BodyText} from '@workday/canvas-kit-react/text';
+import {createStyles} from '@workday/canvas-kit-styling';
+import {system} from '@workday/canvas-tokens-web';
+
+const flexStyles = createStyles({
+  display: 'flex',
+  gap: system.space.x2,
+});
 
 export const WithCount = () => {
   const [text, setText] = React.useState('');
   return (
-    <Box>
-      <Flex gap="xxs">
+    <div>
+      <div className={flexStyles}>
         <Pill onClick={() => setText('The first pill is clicked!')}>
           Shoes
           <Pill.Count>30</Pill.Count>
@@ -16,8 +22,8 @@ export const WithCount = () => {
           Shoes
           <Pill.Count>30</Pill.Count>
         </Pill>
-      </Flex>
+      </div>
       <BodyText size="medium">{text}</BodyText>
-    </Box>
+    </div>
   );
 };
