+++ modules/preview-react/pill/stories/examples/WithReadOnly.tsx
@@ -1,15 +1,20 @@
 import * as React from 'react';
 
 import {Pill} from '@workday/canvas-kit-preview-react/pill';
+import {createStyles} from '@workday/canvas-kit-styling';
+import {system} from '@workday/canvas-tokens-web';
 
-import {Flex} from '@workday/canvas-kit-react/layout';
+const flexStyles = createStyles({
+  display: 'flex',
+  gap: system.space.x2,
+});
 
 export const WithReadOnly = () => (
-  <Flex gap="xxs" id="read-only-list">
+  <div className={flexStyles} id="read-only-list">
     <Pill variant="readOnly">Read-only</Pill>
     <Pill variant="readOnly" maxWidth={250}>
       Read-only but with super long text in case you want to read a paragraph in a Pill which we
       don't recommend
     </Pill>
-  </Flex>
+  </div>
 );
