+++ modules/preview-react/pill/stories/examples/WithRemovable.tsx
@@ -3,14 +3,20 @@ import React from 'react';
 import {Pill} from '@workday/canvas-kit-preview-react/pill';
 // @ts-ignore: Cannot find module error
 import testAvatar from './test-avatar.png';
-import {Box, Flex} from '@workday/canvas-kit-react/layout';
 import {BodyText} from '@workday/canvas-kit-react/text';
+import {system} from '@workday/canvas-tokens-web';
+import {createStyles} from '@workday/canvas-kit-styling';
+
+const flexStyles = createStyles({
+  display: 'flex',
+  gap: system.space.x2,
+});
 
 export const WithRemovable = () => {
   const [text, setText] = React.useState('');
   return (
-    <Box>
-      <Flex gap="xxs">
+    <div>
+      <div className={flexStyles}>
         <Pill variant="removable">
           Pink Shirts
           <Pill.IconButton
@@ -30,8 +36,8 @@ export const WithRemovable = () => {
           <Pill.Label>This is a category that should not exist because it is too long</Pill.Label>
           <Pill.IconButton aria-label="Remove" />
         </Pill>
-      </Flex>
+      </div>
       <BodyText size="medium">{text}</BodyText>
-    </Box>
+    </div>
   );
 };
